Move plotting utils into DataAssistantResult
