+++ great_expectations/rule_based_profiler/types/data_assistant_result/__init__.py
@@ -1,1 +1,2 @@
 from .data_assistant_result import DataAssistantResult
+from .volume_data_assistant_result import VolumeDataAssistantResult
